Diagnostic of RMT
This branch is used to deal with anomolies related to the RMT peripheral.
This particular piece of code detect the problem that level0 and level1 are both 0. It will output a 1us pulse at pin 21 when this happens.
It turns out that this will happen when the filter is enabled and timeout is triggered because we filtered out some signals. Will only happen when the receiver received very weak signal.